[refactor] #67 클래스 도메인 버전 패키지 분리
