+++ src/main/java/com/abcdedu_backend/lecture/entity/v2/AssignmentQuestion.java
@@ -1,5 +1,6 @@
-package com.abcdedu_backend.lecture.entity;
+package com.abcdedu_backend.lecture.entity.v2;
 
+import com.abcdedu_backend.lecture.entity.Assignment;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
