+++ src/main/java/com/abcdedu_backend/lecture/repository/v2/AssignmentAnswerRepository.java
@@ -1,6 +1,6 @@
-package com.abcdedu_backend.lecture.repository;
+package com.abcdedu_backend.lecture.repository.v2;
 
-import com.abcdedu_backend.lecture.entity.AssignmentAnswer;
+import com.abcdedu_backend.lecture.entity.v2.AssignmentAnswer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AssignmentAnswerRepository extends JpaRepository<AssignmentAnswer, Long> {
