+++ src/main/java/com/abcdedu_backend/lecture/service/LectureServiceV2.java
@@ -11,7 +11,14 @@ import com.abcdedu_backend.lecture.dto.request.CreateSubLectureRequest;
 import com.abcdedu_backend.lecture.dto.response.GetAssignmentAnswerResponse;
 import com.abcdedu_backend.lecture.dto.response.GetAssignmentResponseV2;
 import com.abcdedu_backend.lecture.entity.*;
+import com.abcdedu_backend.lecture.entity.v2.AssignmentAnswer;
+import com.abcdedu_backend.lecture.entity.v2.AssignmentAnswerType;
+import com.abcdedu_backend.lecture.entity.v2.AssignmentQuestion;
+import com.abcdedu_backend.lecture.entity.v2.AssignmentSubmission;
 import com.abcdedu_backend.lecture.repository.*;
+import com.abcdedu_backend.lecture.repository.v2.AssignmentAnswerRepository;
+import com.abcdedu_backend.lecture.repository.v2.AssignmentQuestionRepository;
+import com.abcdedu_backend.lecture.repository.v2.AssignmentSubmissionRepository;
 import com.abcdedu_backend.member.entity.Member;
 import com.abcdedu_backend.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
